Implement Python Structures
This implementation, though done half way, would help us a lot.
It's mainly done to avoid the problem occuring when a class
referencing a C object and the C object get's freed before.
With this only renderer and few utils should be written in C everything
should be in Python files.

Add a dependency on attrs, since it's great and lot helpful for us.

Finally, try typing parts of code so that we don't become a blocker when
Manim uses mypy.

Signed-off-by: Naveen M K <naveen@syrusdark.website>